types refactored, extracted to own files
