+++ app/javascript/components/concept-exercise-graph/concept-connection-types.ts
@@ -1,0 +1,21 @@
+export type ConceptConnection = {
+  from: string
+  to: string
+}
+
+export enum ConceptPathState {
+  Unavailable = 'unavailable',
+  Available = 'available',
+  Completed = 'completed',
+}
+
+export type ConceptPathCoordinate = {
+  x: number
+  y: number
+}
+
+export type ConceptPath = {
+  start: ConceptPathCoordinate
+  end: ConceptPathCoordinate
+  state: ConceptPathState
+}
