+++ app/javascript/components/concept-exercise-graph/concept-types.ts
@@ -7,13 +7,6 @@ export enum ConceptState {
   InProgress = 'in_progress',
 }
 
-export interface IExercise {
-  slug: string
-  uuid: string
-  prerequisiteConcepts: string[]
-  unlocksConcepts: string[]
-}
-
 export interface IConcept {
   index: number
   slug: string
@@ -21,19 +14,6 @@ export interface IConcept {
   status: ConceptState
   handleEnter?: MouseEventHandler
   handleLeave?: MouseEventHandler
-}
-
-export type ConceptLayer = string[]
-
-export type ConceptConnection = {
-  from: string
-  to: string
-}
-
-export interface IConceptGraph {
-  concepts: IConcept[]
-  layout: ConceptLayer[]
-  connections: ConceptConnection[]
 }
 
 export function isIConcept(concept: IConcept | undefined): concept is IConcept {
